special_one

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -9,9 +9,9 @@
 
 int main(void)
 {
-	long long int prime;
-	long long int if_prime;
-	long long int NUMBER = 612852475143;
+	unsigned long int prime;
+	unsigned long int if_prime;
+	unsigned long int NUMBER = 612852475143;
 
 	for (prime = NUMBER / 2; prime > 1; prime--)
 	{
@@ -25,7 +25,7 @@ int main(void)
 					break;
 			}
 			if (if_prime == 1)
-				printf("%lld\n", prime);
+				printf("%lu\n", prime);
 			break;
 		}
 	}
